feat: implement attack and randomAttack (#4)
