+++ src/commands/startGame.ts
@@ -3,14 +3,31 @@ import { StartGamePayload } from '../types/types';
 import { getUserById } from '../utils/getUser';
 import { logOutgoing } from '../utils/logging';
 import { WebSocket } from 'ws';
+import { turn } from './turn';
 
 export const startGame = (ws: WebSocket, payload: StartGamePayload) => {
   const { gameId, ships, indexPlayer } = payload;
 
+  const placedShips = ships.map((ship) => {
+    const cells = [];
+
+    for (let i = 0; i < ship.length; i++) {
+      const x = ship.direction ? ship.position.x : ship.position.x + i;
+      const y = ship.direction ? ship.position.y + i : ship.position.y;
+
+      cells.push({ x, y });
+    }
+
+    return {
+      cells,
+      hp: cells.length,
+    };
+  });
+
   if (!GAMES.has(gameId)) {
     GAMES.set(gameId, {
       players: [indexPlayer],
-      ships: { [indexPlayer]: ships },
+      ships: { [indexPlayer]: placedShips },
       isStarted: false,
       currentTurnPlayerId: null,
     });
@@ -21,7 +38,7 @@ export const startGame = (ws: WebSocket, payload: StartGamePayload) => {
       game.players.push(indexPlayer);
     }
 
-    game.ships[indexPlayer] = ships;
+    game.ships[indexPlayer] = placedShips;
 
     const playersReady = game.players.every((playerId) => game.ships[playerId]);
 
@@ -51,45 +68,4 @@ export const startGame = (ws: WebSocket, payload: StartGamePayload) => {
   }
 
   USER_TO_GAME.set(indexPlayer, gameId);
-};
-
-export const turn = (ws: WebSocket, userId: string) => {
-  const gameId = USER_TO_GAME.get(userId);
-  if (!gameId) return;
-
-  const game = GAMES.get(gameId);
-  if (!game) return;
-
-  if (game.currentTurnPlayerId !== userId) {
-    return;
-  }
-
-  const currentIndex = game.players.indexOf(userId);
-  const nextIndex = (currentIndex + 1) % game.players.length;
-  game.currentTurnPlayerId = game.players[nextIndex];
-
-  const response = {
-    type: 'turn',
-    data: JSON.stringify({
-      currentPlayer: game.currentTurnPlayerId,
-    }),
-    id: 0,
-  };
-
-  const json = JSON.stringify(response);
-  logOutgoing(response);
-
-  for (const playerId of game.players) {
-    const player = getUserById(playerId);
-    if (player?.ws.readyState === WebSocket.OPEN) {
-      player.ws.send(json);
-    }
-  }
-};
-
-export const randomAttack = (userId: string) => {
-  console.log('randomAttack userId:', userId);
-  console.log('USER_TO_GAME map:', [...USER_TO_GAME.entries()]);
-  console.log('USER_TO_GAME.has(userId):', USER_TO_GAME.has(userId));
-  console.log('gameId:', USER_TO_GAME.get(userId));
 };
